+++ src/main/java/org/truffleruby/core/array/ArrayOperations.java
@@ -14,10 +14,48 @@ import com.oracle.truffle.api.CompilerDirectives.TruffleBoundary;
 import com.oracle.truffle.api.object.DynamicObject;
 import com.oracle.truffle.api.profiles.ConditionProfile;
 import org.truffleruby.Layouts;
+import org.truffleruby.RubyContext;
+import org.truffleruby.language.objects.shared.SharedObjects;
 
 import java.lang.reflect.Array;
 
 public abstract class ArrayOperations {
+
+    public static boolean isPrimitiveStorage(DynamicObject array) {
+        Object store = getBackingStore(array);
+        return store == null || store instanceof int[] || store instanceof long[] || store instanceof double[];
+    }
+
+    public static boolean verifyStore(DynamicObject array) {
+        final Object backingStore = getBackingStore(array);
+        assert backingStore == null || backingStore instanceof int[] || backingStore instanceof long[] || backingStore instanceof double[] ||
+                backingStore.getClass() == Object[].class : backingStore;
+
+        final RubyContext context = Layouts.MODULE.getFields(Layouts.ARRAY.getLogicalClass(array)).getContext();
+        if (SharedObjects.isShared(context, array)) {
+            final Object store = Layouts.ARRAY.getStore(array);
+
+            if (store != null && store.getClass() == Object[].class) {
+                final Object[] objectArray = (Object[]) store;
+
+                for (Object element : objectArray) {
+                    assert SharedObjects.assertPropagateSharing(context, array, element) : "unshared element in shared Array: " + element;
+                }
+            } else if (store instanceof DelegatedArrayStorage && ((DelegatedArrayStorage) store).hasObjectArrayStorage()) {
+                final DelegatedArrayStorage delegated = (DelegatedArrayStorage) store;
+                final Object[] objectArray = (Object[]) delegated.storage;
+
+                for (int i = delegated.offset; i < delegated.offset + delegated.length; i++) {
+                    final Object element = objectArray[i];
+                    assert SharedObjects.assertPropagateSharing(context, array, element) : "unshared element in shared copy-on-write Array: " + element;
+                }
+            } else {
+                assert isPrimitiveStorage(array);
+            }
+        }
+
+        return true;
+    }
 
     public static int normalizeIndex(int length, int index, ConditionProfile negativeIndexProfile) {
         if (negativeIndexProfile.profile(index < 0)) {
@@ -51,13 +89,18 @@ public abstract class ArrayOperations {
     }
 
     @TruffleBoundary
-    public static Object getBackingStore(DynamicObject array) {
-        return Layouts.ARRAY.getStore(array);
+    private static Object getBackingStore(DynamicObject array) {
+        final Object store = Layouts.ARRAY.getStore(array);
+        if (store instanceof DelegatedArrayStorage) {
+            return ((DelegatedArrayStorage) store).storage;
+        } else {
+            return store;
+        }
     }
 
     @TruffleBoundary
     public static int getStoreCapacity(DynamicObject array) {
-        Object store = getBackingStore(array);
+        Object store = Layouts.ARRAY.getStore(array);
         if (store == null) {
             return 0;
         } else {